Start refactoring to support multi engine peeler.
