+++ tridesclous/peeler_tools.py
@@ -1,0 +1,91 @@
+from collections import OrderedDict, namedtuple
+import numpy as np
+
+
+from .labelcodes import (LABEL_TRASH, LABEL_UNCLASSIFIED, LABEL_ALIEN)
+
+LABEL_LEFT_LIMIT = -11
+LABEL_RIGHT_LIMIT = -12
+LABEL_MAXIMUM_SHIFT = -13
+# good label are >=0
+
+_dtype_spike = [('index', 'int64'), ('cluster_label', 'int64'), ('jitter', 'float64'),]
+
+Spike = namedtuple('Spike', ('index', 'cluster_label', 'jitter'))
+
+
+
+def make_prediction_signals(spikes, dtype, shape, catalogue, safe=True):
+    #~ n_left, peak_width, 
+    
+    prediction = np.zeros(shape, dtype=dtype)
+    for i in range(spikes.size):
+        k = spikes[i]['cluster_label']
+        if k<0: continue
+        
+        #~ cluster_idx = np.nonzero(catalogue['cluster_labels']==k)[0][0]
+        cluster_idx = catalogue['label_to_index'][k]
+        
+        #~ print('make_prediction_signals', 'k', k, 'cluster_idx', cluster_idx)
+        
+        # prediction with no interpolation
+        #~ wf0 = catalogue['centers0'][cluster_idx,:,:]
+        #~ pred = wf0
+        
+        # predict with tailor approximate with derivative
+        #~ wf1 = catalogue['centers1'][cluster_idx,:,:]
+        #~ wf2 = catalogue['centers2'][cluster_idx]
+        #~ pred = wf0 +jitter*wf1 + jitter**2/2*wf2
+        
+        #predict with with precilputed splin
+        r = catalogue['subsample_ratio']
+        pos = spikes[i]['index'] + catalogue['n_left']
+        jitter = spikes[i]['jitter']
+        #TODO debug that sign
+        shift = -int(np.round(jitter))
+        pos = pos + shift
+        
+        #~ if np.abs(jitter)>=0.5:
+            #~ print('strange jitter', jitter)
+        
+        #TODO debug that sign
+        #~ if shift >=1:
+            #~ print('jitter', jitter, 'jitter+shift', jitter+shift, 'shift', shift)
+        #~ int_jitter = int((jitter+shift)*r) + r//2
+        int_jitter = int((jitter+shift)*r) + r//2
+        #~ int_jitter = -int((jitter+shift)*r) + r//2
+        
+        #~ assert int_jitter>=0
+        #~ assert int_jitter<r
+        #TODO this is wrong we should move index first
+        #~ int_jitter = max(int_jitter, 0)
+        #~ int_jitter = min(int_jitter, r-1)
+        
+        pred = catalogue['interp_centers0'][cluster_idx, int_jitter::r, :]
+        #~ print(pred.shape)
+        #~ print(int_jitter, spikes[i]['jitter'])
+        
+        
+        #~ print(prediction[pos:pos+catalogue['peak_width'], :].shape)
+        
+        
+        if pos>=0 and  pos+catalogue['peak_width']<shape[0]:
+            prediction[pos:pos+catalogue['peak_width'], :] += pred
+        else:
+            if not safe:
+                print(spikes)
+                n_left = catalogue['n_left']
+                width = catalogue['peak_width']
+                local_pos = spikes['index'] - np.round(spikes['jitter']).astype('int64') + n_left
+                print(local_pos)
+                #~ spikes['LABEL_LEFT_LIMIT'][(local_pos<0)] = LABEL_LEFT_LIMIT
+                print('LEFT', (local_pos<0))
+                #~ spikes['cluster_label'][(local_pos+width)>=shape[0]] = LABEL_RIGHT_LIMIT
+                print('LABEL_RIGHT_LIMIT', (local_pos+width)>=shape[0])
+                
+                print('i', i)
+                print(dtype, shape, catalogue['n_left'], catalogue['peak_width'], pred.shape)
+                raise(ValueError('Border error {} {} {} {} {}'.format(pos, catalogue['peak_width'], shape, jitter, spikes[i])))
+                
+        
+    return prediction
